Fix USB tracing wrapper on control transfer data
Fix LibusbTracingWrapper to not perform reading beyond end-of-buffer
when dumping control transfer data.

The issue was caused by two related issues: first, we didn't correctly
determine whether a transfer is input or output (because for control
transfers it has to be determined by looking at a different field in the
"setup" packet); second, we incorrectly determined the data length to
dump (because for control transfers the actual data is shorter by the
length of the setup packet).