Fix USB tracing wrapper: cancel completed transfer
Fix LibusbTracingWrapper hitting use-after-free errors when canceling an
already completed (but not destroyed yet) transfer.

The problem was caused by leaving a dangling pointer to the wrapped
transfer in the LibusbTracingWrapper. Before this fix, we were cleaning
the dangling pointer when LibusbFreeTransfer is called, but that left a
time window during which the dangling pointer is still here and can lead
to use-after-free: after the transfer is completed but before it's
freed.