Fix USB tracing wrapper mem leak w/ FREE_TRANSFER (#686)
Fix the LibusbTracingWrapper class to not leak memory when a transfer
with the LIBUSB_TRANSFER_FREE_TRANSFER flag is used.

The leak was caused by the fact that the LibusbTracingWrapper wraps
every transfer into another libusb_transfer before passing it to the
underlying implementation (to LibusbJsProxy). The idea was that the
wrapped transfer has the same parameters as the original transfer,
except for the injected callback that logs additional data. The problem
was that LibusbJsProxy doesn't know anything about the original transfer
object, hence it doesn't perform the automatic deallocation even if it
requested it via the LIBUSB_TRANSFER_FREE_TRANSFER flag.

This bug contributes to the bugfixes tracked by #685. The unit tests
that catch it are in #684.