Fix USB tracing wrapper on unsent transfer cancel (#687)
Fix the LibusbTracingWrapper class to not crash on a debug assertion
when a cancellation of a not-yet-submitted transfer is requested.

The callsites shouldn't normally trigger such calls, but apparently the
original Libusb interface does support them, and we also support that in
the real LibusbJsProxy implementation.

This bugfix contributes to the bugfixes tracked by #685. The
regression unit test is part of #684.