Allow disabling self-reloadon on fatal errors
Normally the Smart Card Connector (and examples like example_cpp App)
self-reload themselves when hitting a fatal error messages in Release
mode. This change allows to suppress this by opening DevTools attached
to the background page and typing this code:

  GoogleSmartCard.disableSelfReloadOnFatalError();

It's intended as a small aid for developers, in situations when they are
investigating a crash happening in the Release mode. It used to be easy
to suppress this by opening DevTools and setting the "goog.DEBUG"
constant manually, however after we enabled advanced JS optimizations
in #420 it became impossible (not only the constant name is minified,
but whole branches of unreachable code are deleted in Release mode now).