Fix Release-mode JS string bugs in example_cpp app (#691)
Add missing quotes for dictionary keys in the
example_cpp_smart_card_client_app code.

They were forgotten when we changed the build scripts to enable advanced
optimizations in Closure Compiler. The advanced optimizations allow the
compiler to shorten dictionary keys in situations when it thinks the
rename is safe (and the compiler doesn't know how to look into usages in
JavaScript code that runs in other pages or in C++ code).