Gracefully handle unexpected revoked PC/SC handles
Remove SCARDCONTEXT/SCARDHANDLE from internal tracking lists when
the PC/SC-Lite core responds with SCARD_E_INVALID_HANDLE.

This shouldn't happen normally, but we already hit bugs in the
PC/SC-Lite that lead to this in the past. In general, we should follow
what the PC/SC-Lite core tells us, and when our data structures diverge
we should fix our structure to follow the PC/SC-Lite's state.