Support building standalone JS lib as ES module
Change //example_js_standalone_smart_card_client_library/ to produce a
new flavor of the library: an ES module that can be import'ed in other
code.

This new flavor provides the same functionality as before, but can be
loaded in environments in which ES modules are preferred (e.g., Chrome
Manifest V3 Extension Service Workers).