Fix JS minification bug in PC/SC Context JS class (#699)
Make sure the addOnDisposeCallback() method is exposed as unminifed
method name on the GoogleSmartCard.PcscLiteClient.Context class.

That way, it can be used by unminifed code, which is one scenario that
we support: the example_js_standalone_client_library that we distribute,
built in Release mode. The regression probably happened in #420, when
we enabled aggressive optimizations in Closure Compiler.