Avoid warning logs on USB device unplugging (#722)
Downgrade a bunch of logs to the INFO level, so that when a smart card
reader is unplugged there are no warnings/errors appearing in the
application log. This is mainly to avoid confusing users who see a red
"Errors" button in chrome://extensions whenever we log a warning/error
and might interpret it as a real problem (which in this case it's not).