Fix application_unittest setup under PNaCl (#723)
Fix crashes in smart_card_connector_app's application unittests due to
attempts to repeatedly remount / and /tmp directories. Before this
commit, only one test was passing and the subsequent ones were crashing
due to the leftover mounts.

Now we're unmounting the directories in each test's teardown. Also we
ignore the unmount() errors because they're fine in these test
scenarios, and in general there's little value to crash on these.