Unit-test automatic PC/SC context cleaning
Add test coverage for auto-cleanup behavior that releases all PC/SC-Lite
SCARDCONTEXTs that were left by a disconnected client.

The test is is written in an suboptimal way as it spins a polling loop
for waiting for an event happening on a background thread; still, even
such a test is better than not having a test at all.