Unit-test SCardStatus with reader/card removal
Add unit test coverage for the scenarios when SCardStatus is called with
a handle to an already removed reader or a reader with an already
removed card.

The tests are implemented in an suboptimal way using a polling loop,
because there's no simple way to be notified about the PC/SC-Lite's
internal state transitions.