Allowlist Epesi Kiosk apps
Add the Kiosk apps from Epesi (in its differnet flavors: production,
development, "acceptance") to the list of trusted client IDs, and
autoallow the permission to talk to smart cards.

This allows, in particular, to use the Smart Card Connector as a
secondary kiosk app (in this mode it has neither any UI nor access to
admin policies).