Workaround std::locale crash in unittest on NaCl
For some reason, constructing std::locale causes a crash on NaCl
toolchain. This leads to failures in the
SmartCardConnectorApplicationSingleClientTest.StringifyError unit test.

Work this around by using an analog from the C library that doesn't
specify the locale. At this point it makes no sense to investigate a
NaCl-specific test-specific issue.