Migrate LogBuffer to ES6 class (#760)
Migrate the //common/js/src/logging/log-buffer.js file and child classes
that depend on it from legacy Closure Compiler classes to ES6 classes.

This is a pure refactoring change. It's is part of the effort to modernize
the code base, as recent Closure Compiler versions stopped
supporting the legacy class syntax.