Migrate ExecutableModule to ES6 class (#759)
Migrate the //common/js/src/executable-module/executable-module.js file
and child classes that depend on it from legacy Closure Compiler classes
to ES6 classes.

This is a pure refactoring change. It's part of the effort to modernize the
code base, as recent Closure Compiler versions stopped supporting the
legacy class syntax.