Fix for inbound USB transfer overflow
When making an inbound USB transfer request to the JavaScript side
(chrome.usb or WebUSB), try to avoid using excessively small transfer
lengths. Specifically, round up the transfer length to the nearest
multiple of the packet size.

The background for this issue is that the JS APIs don't support
timeouts and cancellation properly, which is why it can sometimes
happen that an early API call is resolved with the data that was
logically intended for a later call. If the early call specified a
shorter expected length, it'd fail completely with an error
("babble" in WebUSB or "Inbound transfer overflow" in chrome.usb),
leading to a loss of that data.