Downgrade GetActiveConfigDescriptor warning log
Make following log debug-only instead of a WARNING:

  "LibusbGetActiveConfigDescriptor request failed: No active config
  descriptors were returned by JS API"

This log can be triggered when the USB device disappears suddenly
(e.g., when transitioning to the ChromeOS Lock Screen). This is normal
and doesn't deserve a warning log (which, e.g., results in showing a
red "Errors" button at chrome://extensions).