Fix JS minification issue in C++-to-JS tests (#807)
Integration tests were failing in the pnacl+Release mode, because
Closure Compiler was renaming the test functions from "testSomething"
to "a" or similar. The commit fixes this by putting the function
names as string literals.

Note: This hasn't been noticed, since we don't run tests in "pnacl"
mode on bots, and these integration tests weren't ported to "emscripten"
either.