Add abstract PermissionsChecker interface (#827)
This is a pure refactoring commit. Introduce an abstract interface that
allows to check whether a client (another extension) is allowed to send
commands to Smart Card Connector. Change
PolicyOrPromptingPermissionsChecker to inherit from it. The abstraction
will be used in follow-up changes to mock out the checker in tests.

This is preparation for implementing regression test for #824.