Pass Promise instead of ReadinessTracker (#830)
This is a pure refactoring commit. Change the JS code to pass around
a Promise instead of the whole ReadinessTracker object. This is
sufficient in order to delay command processing until the server is
ready, and is otherwise much simpler to mock out in tests.

This is preparation for implementing regression test for #824.