Make Chromedriver error nonfatal in initialize.sh (#833)
If Chromedriver cannot be downloaded, env/initialize.sh should print the
error message but then proceed to initializing other stuff. Chromedriver
is only used for automated unit-testing, hence it shouldn't be
considered critical for the env initialization.

Also improve the error messages in this part.

This is preparation for fixing #832.