Connector requires ChromeOS >=81 (#839)
Bump the Smart Card Connector's minimum_chrome_version to "81", to
reflect the fact that it cannot be installed onto older ChromeOS
versions (due to the "invalid manifest" error).

The underlying root cause is the "chrome-untrusted://" URL that we've
been using in the manifest.json since 1.3.9.3 (see #838). Unfortunately
we didn't realize the backwards incompatibility issue back then, which
means the current commit won't fix the issue for users on ChromeOS < 81.
Still, it's better to fix the manifest to tell the real requirement now.