Fixup instead of crashing on incorrect ATR size (#841)
Under some circumstances PC/SC-Lite can return incorrect cbAtr,
exceeding the maximum allowed MAX_ATR_SIZE. Fix this up, by treating it
as zero, instead of crashing on an assertion.

The underlying root cause seems to be a race condition between the
thread that handles SCardGetStatusChange() calls and the thread that
manages global structures with reader information. When these read&write
concurrently to the same structure, it can happen that the former reads
the ATR length of the already-uninitialized reader. PC/SC-Lite's
internal convention is that uninitialized reader structures have cbAtr
equal to READER_NOT_INITIALIZED (which is MAX_ATR_SIZE + 1).
Meanwhile we don't fix this race condition in this commit, we at least
mitigate the consequences of the bug by substituting the invalid
integer.

A proper fix will need to be done separately later.