Add PC/SC test running with every simulated reader (#854)
Add a new test for the Smart Card Connector's PC/SC implementation
that's parameterized by the simulated reader type. This lets us test
some basic scenarios work correctly for a range of different devices.

This is a regression test for #849, because one of the simulated devices
used by the new test has the smart card interface number greater than
zero.