[CI] Don't try posting comments in PRs in forks (#858)
Normally our continuous integration scripts post a comment on a Pull
Request once all jobs successfully complete. However, this doesn't
work when the PR is in a fork, as the "Resource not accessible by
integration" permission error occurs
(https://github.com/peter-evans/create-or-update-comment#action-inputs).

This commit makes this job being completely skipped. That way, we avoid
spurious errors for now, until we maybe figure out the right way to do
it.