Fix flaky teardown errors in bridge-integrationtest
bridge-integrationtest.js was occasionally flaky, complaining about
uncatched errors during teardown. It turned out to be because of the
test doing the teardown before the response from C++ was received (in
cases when this response isn't interesting to check).

The fix is to wait until the response arrives before exiting the test.
As drive-by, we also modernize the code in this test via async+await.

This commit contributes to the Emscripten integration test
implementation tracked by #816, because the flakiness was very high in
the Emscripten mode.