[JS-to-C++ test] Toolchain-agnostic build interface (#860)
Move the existing NaCl build rules into a common location, and refactor
them to be an implementation of a toolchain-agnostic makefile rule. As
drive-by, get rid of the hack with double-specified static libraries.

This is preparation for introducing Emscripten build support for the
same tests, and also for new JS-to-C++ integration tests (tracked
by #816).