[JS-to-C++ test] Enable SharedArrayBuffer for Emscripten
Force-enable Chrome's SharedArrayBuffer feature when running integration
tests in the TOOLCHAIN=emscripten mode, because otherwise
multi-threading wouldn't work, which is required by the tests.

The underlying reason is that by default Chrome only allows the
SharedArrayBuffer on pages that have CORS headers set up properly: see
https://developer.chrome.com/blog/enabling-shared-array-buffer/ . Our
tests are run a simple local web server which doesn't have them. As
setting these up would be too tedious, the workaround here is to tweak
Chrome when running the tests instead.

This commit contributes to #816 and is one of the last blockers before
actually enabling integration tests in Emscripten mode and in CI.