[JS-to-C++ test] Rename integration_test_pp_module (#868)
Rename the integration_test_pp_module.cc file to entry_point_nacl.cc,
for consistency with the "entry_point_<toolchain>" scheme. Also move it
outside the google_smart_card_integration_testing/ directory as this
file isn't "public" part of the library.

This is a pure refactoring commit that's related to the #816 effort.