[JS-to-C++ test] Make helper TearDown async
Change IntegrationTestHelper::TearDown() to report its completion
asynchronously via a callback.

This change is similar to the previous change for SetUp(), except that
teardown doesn't need to return any value, and that we have to properly
accoutn for callbacks in the bulk "tear down all helpers" operation.

This commit contributes to #869.