[JS-to-C++ test] Assert helpers torn down at end (#873)
Assert (and crash if it fails) that by the time
IntegrationTestService::Deactivate() is called all previously set up
helpers have been torn down.

This is anyway guaranteed by the current behavior of the test
controller. Also this assumption greatly simplifies transition to the
asynchronous teardown model, because it'll be hard to make
Deactivate() asynchronous as well (it's called in the "entry point"
class destructor).

This commit contributes to #869.