[JS-to-C++ test] Allow developer manual test debug
Let the developer pass the MANUAL_TEST_DEBUGGING=1 environment variable
to make the test runner's web page visible and also to leave it opened
after the tests finish.

We gate this behavior under an environment variable because Continuous
Integration and routine test executions shouldn't mess up with the UI or
hang on failures.

This commit contributes to #869.