Refactor TestingGlobalContext callbacks (#881)
Make a separate callback for each type of the observation: (a) just
message, (b) request message, (c) response message. Previously all three
types were using the same callback type, but the arguments were omitted
or had different meaning depending on the case.

The refactoring should make the code clearer and simplify passing
use-case-specific arguments. This is preparation for USB simulation in
JS-to-C++ tests as tracked by #869.