[JS-to-C++ test] Workaround unhandled promise error (#982)
Workaround spurious "UnhandledPromiseRejection" errors in JS console
happening when a Libusb function failure is simulated.

The root cause seems to be the incompatibility of the goog.Promise's
built-in safety guards and modern EcmaScript's await expression. The
promise, when rejected, thinks that its failure wasn't checked or
propagated, despite that it's what "await" does. I didn't manage to fix
this broken communication, so instead this commit suppresses the broken
safeguard.

This commit contributes to #869, as it's a prerequisite for adding
SCardGetStatusChange() tests.