[makefile] Drop test targets in child makefiles (#1000)
These "test" and corresponding "clean" targets are obsolete everywhere
except the root makefile. The root makefile is more reliable and
scalable, since it ensures dependencies are tracked in a single central
place, they're rebuilt when necessary and in parallel fashion.

This fixes #948, as "make clean" was failing on unusual toolchains like
"coverage" - we never updated those child makefiles to skip irrelevant
toolchains.