Choose latest Chromedriver per Chrome milestone (#1023)
As in practice it's not guaranteed that every public Chrome release gets
a corresponding Chromedriver release with the exact same version, we
change env/initialize.sh to fetch using just milestone instead.

Intended as a quick fix for #1009. It's also a partial revert of #834 (as
Chromedriver maintainers seem to have revived the convenient
LATEST_RELEASE_ URLs).