[chrome-api-jstocxxtest] Add function to switch mocks back to recording (#1008)
Currently all expectations need to be set before any testing takes place, because there's no way to switch between replaying and recording mode in goog.testing framework. Because of that we couldn't factor out repeated pieces of code, such as establishing context, to a separate function.
This CL adds getFreshMock function to allow refactoring in the follow-up CLs.