[style] Import find-files-for-linting Python code
Change the users of the find-files-for-linting.py Python code
(eslient.py and format-code.py) to import it as a Python module rather
than run as a subprocess.

This should avoid unnecessary overhead and data conversions. The change
renames the file to "find_files_for_linting.py" as dashes in file names
aren't Python-friendly.