JS Requester handles exceptions correctly
If sending a message failed, the Requester.postRequest() message should
reject the returned promise instead of failing itself with an exception.

A real example when this can happen is when sending a message that
causes a crash in an Emscripten module, because in that case the
exception is bubbled up synchronously.