JS Requester handles disposal-and-exception-while-sending (#1057)
Improve the JS Requester class to gracefully handle this edge case:
the send() call triggers not only an exception, but also the disposal
of the message channel. The bug was that we were trying to reject the
same request twice (first during disposal, second while handling the
exception).

This is follow-up to the crash handling issue fixed in #823.