[PC/SC tests] Test Connect on all simulated readers
Add tests that execute SCardConnect/SCardDisconnect in pattern tests
that are instantiated for every card+reader combination we simulate
in tests.

We also slightly change the structure of these templated tests, so
that this common functionality is tested for both "regular readers"
as well as for "readers with built-in cards" (e.g., Yubikeys).