[libusb] Handle transfers in FIFO (#1105)
When choosing which of active Libusb transfers should receive a
reply from the JS API, use the FIFO order rather than a random order.
Also change the internal structures so that FIFO is used uniformly,
which we'll use in follow-ups for correctly enqueueing JS API calls.

The FIFO order is more intuitive and better corresponds to Libusb
behavior on other platforms. Moreover, the FIFO property will be
crucial to avoid the situation when an enqueued JS API call is
discarded because the transfer happened to be resolved by the call
originally triggered by a different transfer.

This contributes to #462 and #1104.