[PC/SC test] Avoid deadlock in sim interrupt xfs (#1106)
Move interrupt transfer replies in testing_smart_card_simulation
outside of the mutex.

This is needed for follow-ups that rework how our Libusb port
handles incoming requests, because the new implementation may
synchronously start a new USB API call when receiving the previous
call's result. Without the current change, this would deadlock in
the unit tests.

This change contributes to #1104.