[JS-to-C++ tests] Bulletproof teardowns
Make the code in test cases' tearDown functions resistant to
failures in setUps. This avoids spurious errors like "reading the
dispose property of a null object" after the real test failures.

This commit is a no-op for the successfully built tree, it's only
intended to improve troubleshooting of test failures.