Delete unneeded chrome.runtime.onConnect listener
The Smart Card Connector's background script doesn't currently need to
listen for chrome.runtime.onConnect events.

In the past it was needed for wiring up PC/SC calls that happen within
the application (namely, from the ReaderTracker class), but they were
later migrated to directly send messages to the executable module's
channel.

This cleanup is useful in the context of manifestv3 migration, since
we'll need to open and listen for a bunch of message pipes in various
places, and this old code would've thrown asserts for them.