[mv3] Use Fetcher API in TrustedClientsRegistry
Switch the JSON fetching code from goog.net.XhrIo, which essentially is
XMLHttpRequest, to the Fetcher API.

Not only is the Fetcher API more concise and promise-friendly. The main
reason is that XMLHttpRequest is unsupported in Service Worker contexts,
hence this was one of the blockers for the manifestv3 migration.