[mv3] Collect logs from all pages via messaging
Use message passing for sending logs to the Service Worker (or the
Background Page, when compiled for mv2 Apps) from all other pages of the
extension.

This replaces the mv2-specific mechanism that collected all logs by
directly accessing the background page's LogBuffer variable. This
mechanism was a bit hacky, but had the benefit of being synchronous and
zero-overhead; it's not viable in mv3 Service Workers so has to be
deleted.