[connector] Require ChromeOS 128 or later (#1188)
As we're switching to WebAssembly, adjust the minimum supported ChromeOS
version accordingly.

While in theory >=109 should be the greatest common denominator for both
the Chrome App build as well as Chrome mv3 Extension build, we haven't done
thorough tests on each of those versions, and given the active development
and bugfixing of the WebAssembly technology it'd be unwise to roll out our
wasm-based version to them. Hence we choose the threshold conservatively,
as the ChromeOS Stable version as of the time of writing this.