[pcsc] Compile simclist, tokenparser with PC/SC
Move the compilation of these two mini-libraries from the CCID Driver to
PC/SC-Lite.

Unlike in the upstream *nix implementation where both CCID and
PC/SC-Lite ship with their own version of these mini-libraries, we have
to choose one of the copies since we link everything into a single
WebAssembly executable. Previously we semi-aribtrarily chose to put the
compilation and linking into the CCID Driver's makefile, however this
doesn't work out nicely with the plan to add more drivers.