[pcsc] Let drivers themselves export their methods (#1201)
Add an abstraction that represents a driver, and change PC/SC-Lite to
call it instead of directly referencing CCID's functions.

This is a refactoring that prepares us for introducing multiple drivers
in the application. Note that on *nix PC/SC-Lite achieves the same by
dynamically loading each driver's .so file - something we cannot do in
the webport.