[pcsc] Move known_client_apps.json
Move known_client_apps.json from //third_part/pcsc-lite/ to
//smart_card_connector_app/ and rename it to trusted_clients.json.

This is a purely cosmetic change: one reason is consistency with the
code and more logical layering. Another is that this file has been
mentioned in a bunch of places on Github and emails, and we want to
avoid silently breaking those links while cleaning up the word "nacl"
from source file names.